# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/delay/assign": {
            "post": {
                "description": "assign delay of an order that delivery is passed to an agent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delay"
                ],
                "summary": "assign delay",
                "parameters": [
                    {
                        "description": "necessary item for assign delay",
                        "name": "AssignDelayRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.AssignDelayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.AssignDelayResponse"
                        }
                    },
                    "400": {
                        "description": "some field is invalid",
                        "schema": {
                            "$ref": "#/definitions/controllers.AssignDelayResponse"
                        }
                    },
                    "500": {
                        "description": "other error",
                        "schema": {
                            "$ref": "#/definitions/controllers.AssignDelayResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/delay/report": {
            "get": {
                "description": "report delay of an order that delivery is passed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delay"
                ],
                "summary": "report delay",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReportDelayResponse"
                        }
                    },
                    "400": {
                        "description": "some field is invalid",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReportDelayResponse"
                        }
                    },
                    "500": {
                        "description": "other error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReportDelayResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "report delay of a order that delivery is passed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delay"
                ],
                "summary": "report delay",
                "parameters": [
                    {
                        "description": "necessary item for reporting delay",
                        "name": "DelayReportRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DelayReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.DelayReportResponse"
                        }
                    },
                    "400": {
                        "description": "some field is invalid",
                        "schema": {
                            "$ref": "#/definitions/controllers.DelayReportResponse"
                        }
                    },
                    "500": {
                        "description": "other error",
                        "schema": {
                            "$ref": "#/definitions/controllers.DelayReportResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.AssignDelayRequest": {
            "type": "object",
            "required": [
                "agentId"
            ],
            "properties": {
                "agentId": {
                    "type": "integer"
                }
            }
        },
        "controllers.AssignDelayResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "orderId": {
                    "type": "integer"
                }
            }
        },
        "controllers.DelayReportRequest": {
            "type": "object",
            "required": [
                "orderId"
            ],
            "properties": {
                "orderId": {
                    "type": "integer"
                }
            }
        },
        "controllers.DelayReportResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.ReportDelayResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "result": {}
            }
        }
    }
}